pyproject.toml
[project]
name = "transitwind"
version = "0.1.0"
description = "Personalized Human Design transit interpreter"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pyswisseph>=2.10",
    "sqlalchemy>=2.0",
    "anthropic>=0.49",
    "pyjwt>=2.10",
    "bcrypt>=4.2",
    "python-dotenv>=1.0",
    "pydantic>=2.10",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"
